Move kernel over to git.
We are using the same sources as what is in the tarball. Just that they
are now in git. Copying the kernel tree over to the build dir for now.
Will eventually modify it so that we build from the src directory into
the build directory.

Review URL: http://chromereview.prom.corp.google.com/1177012

git-svn-id: svn://chrome-svn/chromeos/trunk@81 06c00378-0e64-4dae-be16-12b19f9950a1