kernel: move config into kernel git repo
Already moved config into the kernel git repo. This change fixes the
build to use the config which is in the kernel and removes the configs
in the chromeos repo.

Also removed a lot of the kernel version hard-coding that is in the build.

Review URL: http://chromereview.prom.corp.google.com/1188031

git-svn-id: svn://chrome-svn/chromeos/trunk@214 06c00378-0e64-4dae-be16-12b19f9950a1