Remove pybootchartgui since it messes up our repo.
The mirroring process for the internal package repository puts
all mirrored packages into "main" for some reason. Since the Karmic
version of pybootchartgui is in "universe", when we use the internal
repository of packages we need to reference it based on "universe".
However, the external repo will have it at "main". We can't have it
both ways, so we'll need to give up on pybootchartgui until after we
fix up the repo mirroing.

Review URL: http://codereview.chromium.org/509010