Decrease the number of bootstrap packages in install_packages.sh.
We use apt to install the majority of our packages. In order to
get apt running where it can install packages we first need to
force install a small set of packages that have cyclic dependencies.
This CL reduces the required set quite a bit and fixes a bug where
we were installing a few wrong packages than what we were attempting
to install due to a bad glob.

Review URL: http://codereview.chromium.org/519075