A script to manage the whitelist of package maintainer scripts.
The package_script_whitelist.sh is intendended to make it easy
to make passes through the set of packages that have preinst
or postinst maintainer scripts and add them to the whitelist.
The audit mode will use the PAGER to show non-whitelisted preinst
and postinst scripts and ask you if you want to whitelist them,
create a template alternative maintainer script, skip, etc.

There are also commands to add individual scripts or check to
see if they are in the whitelist.

Review URL: http://codereview.chromium.org/543027