Don't build a binary package for autotest because we're not using it. This saves almost 1 minute (out of 6) on my build host for the current full build_autotest.sh. The savings will be more significant once the Chrome tests (200MB+) also get into autotest.
Alternatively, we could still build the binary package and use it as part of
archive_build.sh (i.e., just copy it there instead of creating a new one).
But most people don't run archive_build so it may be better to not create
the binary package as part of build_autotest.sh.

Review URL: http://codereview.chromium.org/661473