A pair of scripts to make it easier to run and debug 32 bit executables on 64 bit systems
For Example:

  # run the 32 bit binary located ad ../platform/foo/a.out
  $ run_32bit.sh ../platform/foo/a.out

  # run a.out from the current directory, and pass it the --verbose=3 command line option
  $ ../../scripts/run_32bit.sh -- a.out --verbose=3

  # debug a.out from the current directory, include ../v8 in the library search path
  $ LIB_PATHS='../v8/' ../../scripts/debug_32bit.sh a.out

run_32bit.sh runs a 32bit executable, debug_32bit.sh starts up gdb so that it's ready to
debug a 32bit target.  Both scripts work from in or out of the chroot.

If you're running from outside the chroot, you can pass a --chroot command line option to
locate the chroot, otherwise a default is computed through the normal mechanism.

Pass the 32 bit target as the parameter to either of these scripts.  If you need to provide
additional parameters to your command, then you should use -- to signify the end of *_32bit.sh
parameters.

These could be useful to people developing parts of chromeos that can be easily isolated
from the rest of the system.  If your code doesn't NEED to run on a chromeos device it
can be a bit easier to try it out on your owm machine during dev/test.

Review URL: http://codereview.chromium.org/1131002