Factory Installer.
Change build_image to support building a factory install image. Also,
a shell script and startup script to run the factory installer.

Change software_update startup script to detect if it's a factory
install and abort if so.

BUG=2378,2379,2380
TEST=Tested factory install worked on device.
Wrapper script to perform factory install at boot.

AU: New arg to install on non-boot device partition. This is used to
install in the factory. Also, switch to shflags for argument parsing.

Review URL: http://codereview.chromium.org/1556022