Only allows mod_image_for_test for default from behavior.
BUG=1970

TEST=Running archive build with clean build and producing
a test image and normal image and zip file.  Also running
with custom from and getting an error code with test_mod=true

This change is required because mod_image_for_test requires it
to be inside chroot.  Since the buildbot is the main user
of this script, this deals with that use case.  Other users
of this script will have to run mod_image_for_test manually
outside of archive_build

Review URL: http://codereview.chromium.org/2008001