cros_workon: modify to be less verbose
We need to be able to use "cros_workon list" in pipelines. Also, would
be nice if the other commands were less verbose. Ideally, they should
die and return a non-zero value.

Also cleaned up the logic in some of the functions. This fixed a couple
of bugs:

1) "cros-workon list" would fail if KEYWORDS_FILE didn't exist
2) "cros-workon start" would fail if KEYWORDS_DIR didn't exit

Change-Id: I76c87f0e36b72c12e8ae937cbadca6cf21a34bad

Review URL: http://codereview.chromium.org/2880005