Update parallel_emerge to use no-env-update feature which skips env_update.
Currently, our build occasionally fails with errors about about symlinks
already existing. This usually only slows down the build, because we have
retries, but it would be good to fix the problem.

This patch updates parallel_emerge to use the no-env-update feature for cross-
compiles. This feature skips the env_update step, which isn't thread-safe, and
isn't necessary for when we're cross-compiling and don't actually need to run
the binaries. Instead we run the env_update at the end, avoiding a thread-safety
issue.

TEST=Ran full build --nousepkg with env_update occuring at end.
BUG=chromium-os:5382

Review URL: http://codereview.chromium.org/3034051