Add --show-output option to parallel_emerge which prints output from jobs.
Also update parallel_emerge to print output in general if you hit
CTRL-C or if a job is running for more than an hour. Also cleanup exit
handling so that emerge exits a bit more cleanly, and so that we don't
hang if emerge throws an Exception.

BUG=none
TEST=Ran ./parallel_emerge --show-output -uDNve chromeos

Review URL: http://codereview.chromium.org/3010056