Update all ebuilds to point to a branch or tag.
Mark all ebuilds in chromiumos-overlay and chromeos-overlay as stable and
point them at the current version we have checked out.

This is useful in case we want to verify that all ebuilds are pointing at the
right commit hash when we are doing a build on a branch. It is intended to be
used prior to tagging and can be used instead of a preflight queue. It is an
alternative to cros_mark_all_as_stable that does not require a chroot and
does not increase the revision number on ebuilds.

Unlike cros_mark_all_as_stable, which only updates the latest stable ebuild
for a given board, this script updates all stable ebuilds for all boards,
including unused stable ebuilds. We do this for the sake of simplicity
because it's hard to know in advance which ebuilds are truly unused, and it
shouldn't hurt to update unused ebuilds.

This script does not support cros_mark_as_stable_blacklist. If there are any
packages in the blacklist, this script exits with an error message.

TEST=Ran this script to point power_manager at branch and tested that we
     built power_manager from the branch.
BUG=chromium-os:7218, chromium-os:6429

Change-Id: Ib2308806c5a4db979b1245a90407fd5b6353e4fd

Review URL: http://codereview.chromium.org/3607014