common.sh: switch out echo -ne for printf and fix incorrect offset
echo -ne isn't portable but printf with octals is so this
makes the switch.

In addition, the current offset is off by 3 and is updating the UUID space.
This was from style cleanup and all functional test passed - of course.
This change fixes that too.

TEST=built and installed and checked ro enforcement works
BUG=chromium-os:7972

Change-Id: Ifb3cb2c6f3219af819baff5750453439e1ba6edf

Review URL: http://codereview.chromium.org/3997001