For the boards already have DB directories, remove the default DB.
The default DB is an all-pass DB. If no DB found, the hardware_Components test
falls back to the all-pass one and skip the real check. In order to prevent
skipping the check, remove the default DB in the boards already have DB
directories.

TEST=run "image_to_usb.sh --factory" and check the default DB removed
BUG=chrome-os-partner:1737

Change-Id: Ia25f2551a5eba3a95fe736e86bfe52c340f329a8

Review URL: http://codereview.chromium.org/5339011