Update prebuilt.py to set permissions correctly on created Packages file.
Since the local Packages file is a temporary file, other users won't have
permissions to read it. rsync copies those permissions over to the remote
server. To ensure remote server has the right permissions, we need to use
chmod a+r on either the local file or the remote file. In this case, I chose
to run rsync on the remote file.

TEST=rsync -av --chmod=a+r foo bar and verify permissions are added to bar.
BUG=chromium-os:5311

Change-Id: Iee4c8df00b36b9f9434546603195ea1906b289f1

Review URL: http://codereview.chromium.org/5501001