Make restart_in_chroot_if_needed run against scripts that are symlinks
This reverts the main change of
84e72fab23c1359785f93685df6a3a45e7612d6e

and undoes the breakage of scripts that are symlinks to
...chromeos/scripts..., but at the cost of re-breaking scripts under
scripts/bin.  The correct fix is a combination of removing all the
symlinks to ...chromeos/scripts... and a more robust way of finding
the chroot path of the executable.  But that will have to wait until
someone else does it next week or I return from vacation

TEST=Ran all of these outside chroot.  None seemed more broken than when run in the chroot.  /setup_board  --board=x86-generic, ./cros_workon list --all --board=x86-mario, ./start_devserver , ./build_image --help, ./build_packages --help, ./cros_upgrade_chroot --help, ./enter_chroot.sh , ./customize_rootfs --help, ./run_remote_tests.sh , ./verify_rootfs_chksum.sh , ./set_shared_user_password.sh
BUG=n0ne

Change-Id: Ic2b7484e53e6a977ed6ae675fe99e109b385020d

Review URL: http://codereview.chromium.org/5946003