parallel_emerge now uninstalls packages correctly when use flags change.
When I added the --force-remote-binary flag, I needed to find a way to
get emerge to install binary packages whose use flags were not fully up
to date. To do this, I removed the --newuse flag from the call to the
emerge depgraph calculation code. This seemed safe at the time because
parallel_emerge already checks the use flags of every package it
installs. Unfortunately, things are a little more complex. Sometimes,
emerge decides to install a different version of a package because of a
use flag change. To implement this, Portage needs to know about the
--newuse flag.

To fix this bug, I've updated parallel_emerge to force in the remote
binary packages using a less intrusive method, that only impacts the
package that is being forced. This method should have less side effects
and allow us to install packages correctly.

TEST=./setup_board --board=tegra2_dev-board --force &&
     ./parallel_emerge --board=tegra2_dev-board -uDNvg opengles &&
     ./parallel_emerge --board=tegra2_dev-board -puDNvg board-devices
     ./parallel_emerge --board=tegra2_dev-board -uDNvg chromeos
--force-remote-binary=chromeos-chrome
BUG=chromium-os:10423

Change-Id: I2057d2992abcb08f184cfaa6c099bd6eb8c21a99

Review URL: http://codereview.chromium.org/6033001