Fix bug 10466: parallel_emerge use flag calculation favors local binhost.
When we merge a package, we need to keep track of the use flags associated with
the package, so that we don't merge a package with the wrong use flags.

In some cases, there might be two copies of the same binary package: One local
copy, and one remote copy. Our current logic for grabbing the use flags favors
the local copy. But Portage itself favors the remote copy. To reconcile this
difference, we just reuse the logic from Portage here.

BUG=chromium-os:10466
TEST=
  1. Run build_packages with a correct binhost
  2. Switch to a binhost with different use flags for some of the packages, but the same version number and package name.
  3. Run EXTRA_BOARD_FLAGS=--emptytree build_packages
  4. Run build_image

Also test #3 with --oldchromebinary.

Change-Id: I9b917d8b8d902e0581d5a5d23ad20940930b114a

Review URL: http://codereview.chromium.org/6055004