Update parallel_emerge to experimentally support Portage 2.1.9.
Portage 2.1.9 changes a few APIs, so parallel_emerge needs to change with it.
With this patch, parallel_emerge basically seems to work with Portage 2.1.9.

Note #1: All of our patches against Portage 2.1.7 will need to be forward
ported to Portage 2.1.9 before we'll have complete support for it. See
http://codereview.chromium.org/6093002/ for that.

Note #2: This includes http://codereview.chromium.org/6055004/ . That patch
can be submitted separately, though, and that'll make this patch smaller.

BUG=chromium-os:10574
TEST=Ran ./build_packages  --board=x86-generic --oldchromebinary with and without new portage.
It completes successfully in both cases.

Review URL: http://codereview.chromium.org/6094001

Change-Id: I0f851b040e266d98933962b4ccc1c85f7851c8ad