Verify that SCRIPTS_DIR exists in build_image.
This hack is necessary for old chroots to ensure that we source a good
common.sh when the one inside the chroot is out of date. We'll be able
to get rid of this after everybody has updated their chroot to have a working
version of common.sh and/or we have deprecated running scripts from the
source directory.

Also fix logic for when common.sh is missing to exit properly.

BUG=chromium-os:11598
TEST=Verify that build_image still works when the old common.sh is present in
     the chroot.

Change-Id: I2e8c2fd5ad06446e669ad665ee2e70f4364958fe

Review URL: http://codereview.chromium.org/6286084