set_shared_user_password.sh updates /etc/shared_user_passwd.txt
This is necessary so that the baselayout ebuild can be
changed to set /etc/passwd and /etc/shadow appropriately.
A previous attempt to do this through an environment
variable in make.conf.user failed because pkg_postinst
from binary packages run in their prebuilt environment
(and, thus, are not picking the user specific variables).

BUG=chromium-os:11664
TEST=ran set_shared_user_password.sh multiple times, observed
/etc/shared_user_passwd.txt and shared_user_passwd.txt getting updated.

Change-Id: I23b77c7417db0fd18b251f7cca62027813b182a8

Review URL: http://codereview.chromium.org/6534003