Enhancements to bootperf, based on recent user feedback.
Enhancement 1:  Fail if we run outside the chroot, because otherwise
you have to wait 10 minutes to find out your results won't be gathered.

Enhancement 2:  Print the location of the log output file, for the
benefits of looky-loos who want to see what's going on.

Change-Id: I388b1855c5b6bb20b0efffbf0babd1f60caf7bcb

BUG=None
TEST=run the script per instructions, and see the new output
TEST=run the script outside the chroot, and see it fail immediately

Review URL: http://codereview.chromium.org/6519039