[crosutils] Make chromium/libcros mismatch messages more semantically accurate
The messages printed out by the code that checks if chromium's libcros and the system's libcros are
in sync were kind of confusing and not quite accurate.  The system library implements a range of API
versions, and it is chromium that requires a particular version of the API.  The messages now clearly
reflect this.

BUG=None
TEST=manual

Change-Id: I7308d7eb724edfb7fa11edbe0fbec86ee12c5afe
Reviewed-on: http://gerrit.chromium.org/gerrit/1136
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>