build_image: incorporate mod_image_for_test
Incorporate mod_image_for_test.sh into build_image to simplify the user
interface to the build system and to remove the redundant setup that
occurs in both scripts.

We're not porting the --force_copy, --inplace, or --yes flags, because
the user is no longer passing in an existing image as input.

BUG=chromium-os:12899
TEST=Run build_image w/ various flag combos, inspect output dir, and
  boot resulting images. Flags tested include: none, --test, --factory,
  --factory_install, --factory_install --test.

Change-Id: Ie3b504b3c58fbe19f1ce351985e5f1043353766a
Reviewed-on: http://gerrit.chromium.org/gerrit/2421
Tested-by: Vince Laviano <vlaviano@chromium.org>
Reviewed-by: Anush Elangovan <anush@chromium.org>